fix bug

+++ newopus/LTPAnalysisFilter.go
@@ -32,7 +32,7 @@ func silk_LTP_analysis_filter(
 			LTP_res_ptri := LTP_res_ptr + i
 			LTP_res[LTP_res_ptri] = x[x_ptr2+i]
 
-			LTP_est = silk_SMULBB(x[x_lag_ptr+SilkConstants.LTP_ORDER/2], Btmp_Q14[0])
+			LTP_est = silk_SMULBB(int(x[x_lag_ptr+SilkConstants.LTP_ORDER/2]), int(Btmp_Q14[0]))
 			LTP_est = silk_SMLABB_ovflw(LTP_est, int(x[x_lag_ptr+1]), int(Btmp_Q14[1]))
 			LTP_est = silk_SMLABB_ovflw(LTP_est, int(x[x_lag_ptr]), int(Btmp_Q14[2]))
 			LTP_est = silk_SMLABB_ovflw(LTP_est, int(x[x_lag_ptr-1]), int(Btmp_Q14[3]))
@@ -41,10 +41,10 @@ func silk_LTP_analysis_filter(
 			LTP_est = silk_RSHIFT_ROUND(LTP_est, 14)
 
 			tmp := int(x[x_ptr2+i]) - LTP_est
-			LTP_res[LTP_res_ptri] = silk_SAT16(tmp)
+			LTP_res[LTP_res_ptri] = int16(silk_SAT16(tmp))
 
 			gain := int(invGains_Q16[k])
-			smulwb_result := silk_SMULWB(gain, LTP_res[LTP_res_ptri])
+			smulwb_result := silk_SMULWB(gain, int(LTP_res[LTP_res_ptri]))
 			LTP_res[LTP_res_ptri] = int16(smulwb_result)
 
 			x_lag_ptr++
